Added the sums of macros (calories, proteins, fats, carbs) for a specific meal in the meals menu.

+++ app/src/main/java/com/example/projectgym/user.kt
@@ -1,7 +1,5 @@
 package com.example.projectgym
 
-import kotlinx.datetime.DayOfWeek
-
 class User (
     var name: String,
     var age: Float,
@@ -41,7 +39,7 @@ class ListMeals (
     }
 }
 
-class Meals (
+class Meals(
     var name: String,
     var products: MutableList<Products> = mutableListOf()
 ) {
@@ -50,11 +48,11 @@ class Meals (
     }
 }
 
-class Products (
+class Products(
     var name: String,
-    var calories: Float,
-    var proteins: Float,
-    var fats: Float,
-    var carbs: Float
+    var calories: Double,
+    var proteins: Double,
+    var fats: Double,
+    var carbs: Double
 )
